Move indexer to class, remove `store` references from said class for consistency
