+++ src/Commands/ElasticsearchIndexCommand.php
@@ -2,19 +2,22 @@
 
 namespace Rapidez\Core\Commands;
 
-use Carbon\Carbon;
 use Exception;
-use Illuminate\Support\Arr;
+use Illuminate\Console\Command;
 use Rapidez\Core\Facades\Rapidez;
-use Rapidez\Core\Jobs\IndexJob;
 use Rapidez\Core\Models\Store;
 
-abstract class ElasticsearchIndexCommand extends InteractsWithElasticsearchCommand
+abstract class ElasticsearchIndexCommand extends Command
 {
-    public string $alias;
-    public string $index;
+    public ElasticsearchIndexer $indexer;
 
-    public function indexAllStores(string $indexName, callable|iterable $items, callable|array $mapping, callable|string $id = 'id')
+    public function __construct(ElasticsearchIndexer $indexer)
+    {
+        parent::__construct();
+        $this->indexer = $indexer;
+    }
+
+    public function indexAllStores(string $indexName, callable|iterable $items, callable|array $mapping, callable|string $id = 'id'): void
     {
         $stores = Rapidez::getStores();
         foreach ($stores as $store) {
@@ -22,62 +25,28 @@ abstract class ElasticsearchIndexCommand extends InteractsWithElasticsearchComma
         }
     }
 
-    public function indexStore(Store $store, string $indexName, callable|iterable $items, callable|array $mapping, callable|string $id = 'id')
+    public function indexStore(Store $store, string $indexName, callable|iterable $items, callable|array $mapping, callable|string $id = 'id'): void
     {
-        $this->line('Indexing `'.$indexName.'` for store: '.$store->name);
-
+        $this->line('Indexing `'.$indexName.'` for store '.$store->name);
         try {
-            $this->prepareIndex($store, $indexName);
-            $this->indexItems($items, $mapping, $id);
-            $this->finishIndex();
+            $this->prepareIndexerWithStore($store, $indexName);
+            $this->indexer->index($this->dataFrom($items), $mapping, $id);
+            $this->indexer->finish();
         } catch (Exception $e) {
-            $this->abortIndex();
+            $this->indexer->abort();
 
             throw $e;
         }
     }
 
-    public function indexItems(callable|iterable $items, callable|array $mapping, callable|string $id = 'id')
-    {
-        $currentData = value($items, config()->get('rapidez.store_code'));
-        foreach ($currentData as $item) {
-            $this->indexItem($item, $mapping, $id);
-        }
-    }
-
-    public function indexItem(object $item, callable|array $mapping, callable|string $id = 'id')
-    {
-        $currentValues = is_callable($mapping)
-            ? $mapping($item)
-            : Arr::only((array) $item, $mapping);
-
-        $currentId = is_callable($id)
-            ? $id($item)
-            : $item[$id];
-
-        IndexJob::dispatch($this->index, $currentId, $currentValues);
-    }
-
-    public function finishIndex()
-    {
-        $this->switchAlias($this->alias, $this->index);
-    }
-
-    public function abortIndex()
-    {
-        $this->deleteIndex($this->index);
-    }
-
-    public function createAlias($store, $indexName)
-    {
-        $this->alias = config('rapidez.es_prefix').'_'.$indexName.'_'.$store->store_id;
-        $this->index = $this->alias.'_'.Carbon::now()->format('YmdHis');
-    }
-
-    public function prepareIndex($store, $indexName, array $mapping = [], array $settings = [])
+    public function prepareIndexerWithStore(Store $store, string $indexName, array $mapping = [], array $settings = []): void
     {
         Rapidez::setStore($store);
-        $this->createAlias($store, $indexName);
-        $this->createIndex($this->index, $mapping, $settings);
+        $this->indexer->prepare(config('rapidez.es_prefix').'_'.$indexName.'_'.$store->store_id, $mapping, $settings);
+    }
+
+    public function dataFrom(callable|iterable $items)
+    {
+        return value($items, config()->get('rapidez.store_code'));
     }
 }
